requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.0.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
cryptography==45.0.4
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
Django==5.1.5
django-admin-two-factor==0.1.1
django-axes==8.0.0
django-colorfield==0.12.0
django-cors-headers==4.6.0
django-extensions==4.1
django-formtools==2.5.1
django-ipware==7.0.1
django-json-widget==2.0.1
django-material==1.12.0
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-storages==1.14.4
django-tenant-schemas==1.12.0
django-tenants==3.7.0
django-two-factor-auth==1.17.0
django-unfold==0.45.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==37.1.0
fastapi==0.115.6
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
gradio==5.11.0
gradio_client==1.5.3
groq==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
hyperframe==5.2.0
idna==3.10
ifcopenshell==0.8.0
isodate==0.6.1
Jinja2==3.1.5
jiter==0.8.2
keras==2.11.0
lark==1.2.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.1
oauthlib==3.3.1
openai==1.59.5
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
phonenumbers==9.0.6
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.1
pypng==0.20220715.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-ipware==3.0.0
python-multipart==0.0.20
python-telegram-bot==20.6
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.2.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
ruff==0.8.6
safehttpx==0.1.6
safetensors==0.5.2
schedule==1.2.2
semantic-version==2.10.0
sendgrid==6.11.0
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
soupsieve==2.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
tweepy==4.16.0
typer==0.15.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
youtube-transcript-api==1.2.1
